Apply viewerScript changes to js generator
