+++ packages/create-yoshi-app/templates/platform-app/src/platform/exampleWidgetController.js
@@ -1,21 +1,23 @@
-/* eslint-disable no-unused-vars */
-export function exampleWidgetControllerFactory(
-  {
-    appParams,
-    setProps,
-    $w,
-    wixCodeApi,
-    platformApi,
-    config,
-    compId,
-    connections,
-    livePreviewMode,
-    platformAPIs,
-    type,
-    warmupData,
-  },
-  { experiments, locale },
-) {
+import { EXPERIMENTS_SCOPE } from '../config/index';
+import Experiments from '@wix/wix-experiments';
+
+function getLocale({ wixCodeApi }){
+  return wixCodeApi.window.locale || 'en';
+}
+
+async function getExperimentsByScope(scope) {
+  const experiments = new Experiments({
+    scope,
+  });
+  await experiments.ready();
+  return experiments.all();
+}
+
+export async function exampleWidgetControllerFactory(controllerConfig){
+  const { appParams, setProps } = controllerConfig;
+  const locale = getLocale(controllerConfig);
+  const experiments = await getExperimentsByScope(EXPERIMENTS_SCOPE);
+
   return {
     pageReady() {
       setProps({
