+++ packages/create-yoshi-app/templates/platform-app/src/platform/exampleWidgetController.spec.js
@@ -1,7 +1,18 @@
 import { exampleWidgetControllerFactory } from './exampleWidgetController';
+import LaboratoryTestkit from '@wix/wix-experiments/dist/src/laboratory-testkit';
+import { EXPERIMENTS_SCOPE } from '../config';
+
+export function mockExperiments(scope, experiments) {
+  new LaboratoryTestkit()
+    .withScope(scope)
+    .withBaseUrl(window.location.href)
+    .withExperiments(experiments)
+    .start();
+}
 
 describe('exampleWidgetControllerFactory', () => {
   it('should call setProps with data', async () => {
+    mockExperiments(EXPERIMENTS_SCOPE, { someExperiment: 'true' });
     const setPropsSpy = jest.fn();
     const appParams = {
       baseUrls: {
@@ -11,18 +22,17 @@ describe('exampleWidgetControllerFactory', () => {
     const locale = 'locale';
     const experiments = { someExperiment: 'true' };
 
-    const controller = exampleWidgetControllerFactory(
-      {
-        appParams,
-        setProps: setPropsSpy,
+    const controller = await exampleWidgetControllerFactory({
+      appParams,
+      setProps: setPropsSpy,
+      wixCodeApi: {
+        window: {
+          locale,
+        },
       },
-      {
-        experiments,
-        locale,
-      },
-    );
+    });
 
-    await controller.pageReady();
+    controller.pageReady();
 
     expect(setPropsSpy).toBeCalledWith({
       name: 'World',
