+++ packages/create-yoshi-app/templates/platform-app/src/platform/viewerScript.js
@@ -1,31 +1,7 @@
-import { EXPERIMENTS_SCOPE } from '../config';
-import { exampleWidgetControllerFactory } from './exampleWidgetController';
-import Experiments from '@wix/wix-experiments';
-
-function getLocale({ wixCodeApi }) {
-  return wixCodeApi.window.locale || 'en';
-}
-
-async function getExperimentsByScope(scope) {
-  const experiments = new Experiments({
-    scope,
-  });
-  await experiments.ready();
-  return experiments.all();
-}
+import { exampleWidgetControllerFactory} from './exampleWidgetController';
 
 function createControllers(controllersConfig) {
-  const controllerConfig = controllersConfig[0];
-  const locale = getLocale(controllerConfig);
-
-  return [
-    getExperimentsByScope(EXPERIMENTS_SCOPE).then(experiments =>
-      exampleWidgetControllerFactory(controllerConfig, {
-        experiments,
-        locale,
-      }),
-    ),
-  ];
+  return [exampleWidgetControllerFactory(controllersConfig[0])];
 }
 
 export const viewerScript = {
